Standardized partner privacy UI, removed community settings, added data export/delete, updated date formatting to Azerbaijani long month names, and integrated partner-privacy route into index. Also prepared date-utils translations and adjusted shopping list timestamps to display added times.

X-Lovable-Edit-ID: edt-e408ab1a-53c3-40bb-a6bd-ff5eb0eca4f2

+++ src/components/tools/WeightTracker.tsx
@@ -242,7 +242,7 @@ const WeightTracker = forwardRef<HTMLDivElement, WeightTrackerProps>(({ onBack }
             <div className="flex justify-between mt-2">
               {entries.slice(0, 7).reverse().map((entry) => (
                 <span key={entry.id} className="text-xs text-muted-foreground">
-                  {new Date(entry.entry_date).toLocaleDateString('az-AZ', { day: 'numeric', month: 'short' })}
+                  {new Date(entry.entry_date).toLocaleDateString('az-AZ', { day: 'numeric', month: 'long' })}
                 </span>
               ))}
             </div>
